Move pre-mapping code to utils
