+++ listenbrainz_spark/recommendations/create_dataframes.py
@@ -54,14 +54,11 @@ from data.model.user_missing_musicbrainz_data import UserMissingMusicBrainzDataR
 from data.model.user_cf_recommendations_recording_message import (UserCreateDataframesMessage,
                                                                   UserMissingMusicBrainzDataMessage)
 
-import unidecode
 from flask import current_app
-from pyspark.sql.types import StringType
 import pyspark.sql.functions as func
 from pyspark.sql import Row
 from pyspark.sql.window import Window
-from pyspark.sql.functions import rank, col, row_number, udf
-
+from pyspark.sql.functions import rank, col, row_number
 
 
 # Some useful dataframe fields/columns.
@@ -125,73 +122,6 @@ def generate_dataframe_id(metadata):
     """ Generate dataframe id.
     """
     metadata['dataframe_id'] = '{}-{}'.format(config.DATAFRAME_ID_PREFIX, uuid.uuid4())
-
-
-def get_unique_rows_from_mapping(df):
-    """ Get unique rows from the mapping.
-    """
-    msid_mbid_mapping_df = df.select('mb_artist_credit_id',
-                                     'mb_artist_credit_mbids',
-                                     'mb_artist_credit_name',
-                                     'mb_recording_mbid',
-                                     'mb_recording_name',
-                                     'mb_release_mbid',
-                                     'mb_release_name',
-                                     'msb_artist_credit_name_matchable',
-                                     'msb_recording_name_matchable') \
-                             .distinct()
-
-    return msid_mbid_mapping_df
-
-
-def unaccent_artist_and_track_name(df):
-    """ Remove accents from artist and track name.
-
-        Args:
-            df: Dataframe to process.
-
-        Returns:
-            res_df: Dataframe with unaccented artist name and track name.
-    """
-
-    def get_unaccented_string(accented_string):
-        return unidecode.unidecode(accented_string)
-
-    unaccent_udf = udf(get_unaccented_string, StringType())
-
-    intermediate_df = df.withColumn("unaccented_artist_name", unaccent_udf(df.artist_name))
-
-    res_df = intermediate_df.withColumn("unaccented_track_name", unaccent_udf(intermediate_df.track_name))
-
-    return res_df
-
-
-def convert_text_fields_to_matchable(df):
-    """ Convert text fields (names i.e artist_name, track_name etc) to matchable field.
-        The following steps convert a text field to a matchable field:
-            1.  Unaccent the text.
-            2.  Remove punctuations and whitespaces.
-            3.  Convert to lowercase.
-
-        Args:
-            df: Dataframe to process
-
-        Returns:
-            res_df: Dataframe with artist_name and track_name converted to matchable fields.
-    """
-    unaccent_df = unaccent_artist_and_track_name(df)
-
-    intermediate_df = unaccent_df.withColumn(
-        "artist_name_matchable",
-        func.lower(func.regexp_replace("unaccented_artist_name", '[^A-Za-z0-9]+', ""))
-    )
-
-    res_df = intermediate_df.withColumn(
-        "track_name_matchable",
-        func.lower(func.regexp_replace("unaccented_track_name", '[^A-Za-z0-9]+', ""))
-    )
-
-    return res_df
 
 
 def save_dataframe(df, dest_path):
@@ -268,7 +198,7 @@ def get_listens_for_training_model_window(to_date, from_date, metadata, dest_pat
         raise
 
     partial_listens_df = utils.get_listens_without_artist_and_recording_mbids(training_df)
-    return convert_text_fields_to_matchable(partial_listens_df)
+    return utils.convert_text_fields_to_matchable(partial_listens_df)
 
 
 def get_data_missing_from_musicbrainz(partial_listens_df, msid_mbid_mapping_df):
@@ -314,10 +244,10 @@ def get_data_missing_from_musicbrainz(partial_listens_df, msid_mbid_mapping_df):
                                               'release_name',
                                               'track_name',
                                               'user_name') \
-                                     .agg(func.max('listened_at').alias('listened_at')) \
-                                     .withColumn('rank', row_number().over(window)) \
-                                     .where(col('rank') <= 200) \
-                                     .toLocalIterator()
+        .agg(func.max('listened_at').alias('listened_at')) \
+        .withColumn('rank', row_number().over(window)) \
+        .where(col('rank') <= 200) \
+        .toLocalIterator()
 
     return missing_musicbrainz_data_itr
 
@@ -518,7 +448,7 @@ def main(train_model_window=None):
 
     current_app.logger.info('Loading mapping from HDFS...')
     df = utils.read_files_from_HDFS(path.MBID_MSID_MAPPING)
-    msid_mbid_mapping_df = get_unique_rows_from_mapping(df)
+    msid_mbid_mapping_df = utils.get_unique_rows_from_mapping(df)
     current_app.logger.info('Number of distinct rows in the mapping: {}'.format(msid_mbid_mapping_df.count()))
 
     current_app.logger.info('Mapping listens...')
